refactor(test): P5-2 follow-up — extract Canvas2D shim post-Wave-2 cluster
Continue test_canvas2d_shim.cpp split. Halves the file by extracting
the 1,487-line post-Wave-2 coverage into a sibling TU.

Bundled clusters (all under different pulp issue numbers but unified
by "tests for shim entries that landed after the original #964
coverage sweep"):
  * #1525 — fillText / strokeText maxWidth + glyph cluster handling
  * #1521 — arc-as-path (DIVERGE → PASS)
  * #1520 — ctx.direction / ctx.filter
  * #1526 — catalog hygiene round-trip for already-supported props
  * Wave 4 c2d cleanup — lineDashOffset re-flush on assignment
  * #1527 — getTransform / resetTransform + isPointInPath /
    isPointInStroke

* test_canvas2d_shim.cpp: 3,121 → 1,634 lines (-1,487, -47.6%)
* New target pulp-test-canvas2d-shim-late: 1,571 lines, 51 TEST_CASEs,
  145 assertions.

The run_in_bridge + ScriptedBridge helpers are duplicated locally per
the established extracted-TU pattern.
